ingestion/requirements.txt
asyncpg==0.29.0
psycopg2-binary==2.9.9
pgvector==0.3.2
sqlalchemy[asyncio]==2.0.30
tree-sitter==0.21.3
tree-sitter-languages==1.10.2
gitpython==3.1.43
voyageai==0.2.3
anthropic==0.28.0
python-dotenv==1.0.1
typer==0.12.3
rich==13.7.1
tiktoken==0.7.0
